Update README.md
This project was bootstrapped with Create React App.

Clone the repository into your local machine using git clone https://github.com/ayomideEnoch/google-search-homepage.

Ensure to run npm install or yarn install on your terminal once you have access to this folder on your code editor. This is because the node_modules folder (which is the heaviest folder created using create-react-app has numerous files estimated to be about 25,000 to 30,000 in number as at the time Ayomide Ibosiola wrote this code) is hidden in a .gitignore file placed in the root directory of the project when the code was pushed to GitHub. A gitignore file specifies intentionally untracked files that Git should ignore. GitHub avoids the pushing of such heavy folder to the repository and deals with it perfectly with the npm install / yarn install command. When this is done, all the dependencies required for the project to run are downloaded in a newly created node_modules folder, a package.json file and a package-lock.json file.

Run npm start or yarn start on your ternminal afterwards depending on the package manager you are using.

Congratulations! 
There you have the Google search homepage seating on your PC screen. 
The Google Search button redirects to https://google.com/search?q={users-text-here} when clicked on while the I'm Feeling Lucky button leads to Perfects Engineering homepage.